Graphic Symbols font
MEGA symbol for use as a macro key, like the megakey macro (that's a bad name now)
For use in tables to explain shifted codes.
It is not for listings. The mega80/40.ttf font is used in those cases.

Also moved the element-catalogue.tex into it's own file.

Font demo added to the catalogue.