remove slight ambiguity from configuring sd card
After a discord discussion it was noted that the IMPORTANT message does a good job of conveying that the MEGA65 comes with an SDCARD, but you may need to prepare a new one at times. But then the document said depending on the model you may or may not have received an sd card.
This is a proposed reword to not conflict with the original message that you DO receive an sdcard when you purchase the MEGA65.

I was unsure if the next statement about not doing it on a PC should be brought into this same paragraph or not. 

Please feel free to reformat or change as seen fit.