Remove "variable$ =" from MID$'s format.
The format of other functions do not have their format's in assignment context. 

This is a little different, because MID$ can be used on the LHS of =. But I still think it's a mistake to have this because it implies that MID$ can only be used in assignment context, and not as part of any string expression.