Memory doc changes for release v0.96  (#548)
This updates the chapters and appendices related to the memory system with changes intended for release v0.96, and other corrections.

We're merging to master earlier than the release because it generally improves the state of the documentation, and this is unlikely to affect anyone currently using the documentation in a meaningful way. Through release testing we know that essentially no existing programs are affected by the core changes.