New version of BASIC preamble
I suspect that in the future we will move the BASIC material from the User's Guide to a dedicated BASIC 65 book. Until then, I don't want to reverse the original decision to include this material here. So I'm improving it in place for now.

This also normalizes a few BASIC reference labels.