2nd edition revision (#551)
This revises the User's Guide in preparation for the v0.96 release and new printing for batch #3 delivery.

Goals for this revision:
* Focus on the MEGA65 production model.
* Focus on new users.
* Update procedures and screenshots to represent the most recent best practices and v0.96 behaviors.
* Enhance information about upgrading the platform components, including how to register with Filehost and get the latest release package.
* Extend information about using disks and disk images.
* Give new users enough information to start finding and downloading software and loading it onto their computer.

This also includes an editing pass over the BASIC reference, mostly style consistency and minor factual corrections.

Still to-do: information on features still in progress, including slot 0 flashing, Ethernet file transfer, and M65Connect. LaTeX code style issues were identified in code review that will be addressed in separate commits, so they're not squashed with this one. In particular, I want to address line wrap style with an auto-format solution that needs testing to make sure it doesn't break our listing environments.