Restore CURSOR ON/OFF in the BASIC reference.
I removed this in the 2nd ed. revision because I swear someone had said it had been removed from a previous ROM a while ago. I can't find documentation on what convinced me of this. It's still in the ROM code and seems to work fine, so I have restored the documentation for it for now.