Update PALETTE example to clear screen and use correct paint command (#617)
After `SCREEN SET 1,1` the screen is now cleared with `SCREEN CLR 0`. Alternatively the `SCNCLR` command could have been used. Without this the screen has random noise when this example is run.

Also, I think the `PAINT` command was being used incorrectly. It had 4 parameters, but the 4th parameter is only used when the 3rd parameter is set to `1`, which it wasn't. The 3rd parameter was set to `0`, which appears to be the default and can be omitted. To simplify, the last two parameters were truncated. This does not change how the example works, but seemed like an improper usage of the command.