Removing HWERRATA level 2 from documentation
Nobody is using this and it is being removed from the implementation, possibly to be reconsidered as a non-errata optional feature.

"SD Card Busy Flash behavior" was added to the docs but is not yet implemented. I'll keep the entry for now but document it as unimplemented, in case this gets merged into the core soon.