Fix page numbering so "introduction" chapters are part of the body text, not front matter.
We can consider moving this to front matter with roman numeral page numbers, but before this change it was just restarting from arabic page number 1 twice, which is wrong. The User Guide Introduction chapter is assigned a chapter number of 1, so it makes sense to make it part of the main book, at least until we do something special with chapter numbering.